Initial Commit
This role supports the setup of mailman and can configure postfix and
nginx for you. I used Mark Saprios "Mailman 3 installation experience"
article from the Mailman Wiki[0] as a small guide while writing this
role. I noticed some differences or things that did not quite work out.
Also, instead of using a traditional cronjob, this role solely relies on
systemd services and timers as a more modern approach.

This role has NOT been thoroughly tested, but seems to work flawlessly
on my simple Debian 10 based test setup. There are many thing that need
to be done before this role can be published to ansible galaxy. Besides
testing on a multitude of platforms (the platforms in meta/main.yml are
guesses based on my experience with those platforms in the past) I have
not checked all of the variables. In the past, I've also made some
syntax errors by using the colon syntax for variable declaration in
python files such as the settings_local.py. Whilst I have checked it
once over, it may be possible that some of those yaml-syntax traces are
still in python files. Before this role can be released, documentation
must also be written.

Feel free to test this role out and provide feedback. If you have
questions on how to use and test this role, want to provide feedback or
even patches (awesome) you can either send a E-Mail directly to me
(hello [at] e1mo [dot] de) or preferably to the mailing list at
`~e1mo/ansible-role-mailman [at] lists.sr.ht`[1]. The issue tracker is
also located at sourcehut[2].

I would like to thank multiple people for their help during the creation
of this role:

- Mark Saprio for his original guide[0] which was a helpful guide during
  the creation of this role
- The entire mailman team and all contributors for their work and,
  especially for the mailman core, the extensive documentation[3]
- Countless stackoverflow users and multiple helpful entries in the
  mailman wiki[4] and multiple posts on the mailman-developers list for
  their small hints[5] that came up while searching the web for obscure
  errors
- n0emis[6] for the original idea for writing this role and em0lar[7]
  for their help in testing the mailman installed using this role.

[0]: https://wiki.list.org/DOC/Mailman%203%20installation%20experience
[1]: https://lists.sr.ht/~e1mo/ansible-role-mailman
[2]: https://todo.sr.ht/~e1mo/ansible-role-mailman
[3]: https://docs.mailman3.org/en/latest/
[4]: https://wiki.list.org
[5]: https://mail.python.org/mailman3/lists/mailman-developers.python.org/
[6]: https://www.n0emis.eu
[7]: https://www.em0lar.de