Add role Documentation in README
We now have a documentation that describes all available variables and
also provides some information about limitations of this role. Further
more, the `defaults/main.yml` was restructured a bit to group and
prioritize by component and how important it is to change those values.

One ore more example Playbook(s) are still needed, especially for others
to test out this role.